Qdrant_doc_ingestion/requirements.txt
qdrant-client 
sentence-transformers
biopython